Move the editor into its own module
