+++ .gitignore
@@ -41,23 +41,12 @@ playground.xcworkspace
 .build/
 .swiftpm
 
+## Production Build Products
+/Demo-Android/Gutenberg/src/main/assets/assets
+/Demo-Android/Gutenberg/src/main/assets/index.html
 
-## CocoaPods
-#
-# We recommend against adding the Pods directory to your .gitignore. However
-# you should judge for yourself, the pros and cons are mentioned at:
-# https://guides.cocoapods.org/using/using-cocoapods.html#should-i-check-the-pods-directory-into-source-control
-#
-
-Pods/
-
-
-## Carthage
-#
-# Add this line if you want to avoid checking in source code from Carthage dependencies.
-
-Carthage
-
+/Sources/GutenbergKit/Gutenberg/assets
+/Sources/GutenbergKit/Gutenberg/index.html
 
 # Logs
 logs
@@ -95,9 +84,6 @@ bower_components
 
 # node-waf configuration
 .lock-wscript
-
-# Compiled binary addons (https://nodejs.org/api/addons.html)
-build/Release
 
 # Dependency directories
 node_modules/
