+++ Demo-Android/app/src/main/java/com/example/gutenbergkit/MainActivity.kt
@@ -7,6 +7,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import org.wordpress.gutenberg.GutenbergView
 
 
 class MainActivity : AppCompatActivity() {
