+++ Demo-Android/gradle/libs.versions.toml
@@ -25,4 +25,5 @@ androidx-webkit = { group = "androidx.webkit", name = "webkit", version.ref = "w
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
 jetbrains-kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
+android-library = { id = "com.android.library", version.ref = "agp" }
