+++ Makefile
@@ -6,9 +6,9 @@ build: npm-dependencies
 	npm --prefix ReactApp/ run build
 
 	# Copy build products into place
-	rm -rf ./Sources/GutenbergKit/Gutenberg/ ./Demo-Android/app/src/main/assets/
+	rm -rf ./Sources/GutenbergKit/Gutenberg/ ./Demo-Android/Gutenberg/src/main/assets/
 	cp -r ./ReactApp/dist/. ./Sources/GutenbergKit/Gutenberg/
-	cp -r ./ReactApp/dist/. ./Demo-Android/app/src/main/assets/
+	cp -r ./ReactApp/dist/. ./Demo-Android/Gutenberg/src/main/assets
 
 dev-server: npm-dependencies
 	npm --prefix ReactApp/ run dev
